+++ crates/db/sqlite/src/tests/mod.rs
@@ -10,6 +10,7 @@
 
 #![allow(clippy::arithmetic_side_effects, clippy::indexing_slicing)]
 
+mod balance;
 mod bytes;
 mod constraint;
 mod date;